OCR/api/requirements.txt.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.3.0
pytesseract==0.3.10
Werkzeug==3.0.3